feat: k8s CRDs & integration tests
* refactor/fix CRDs
* test crds on CI
* testutil: manifest generators
* types: use uint16 for ports
* TODO: `provider/cluster/kube` k8s-integration tests

fixes #631
refs  #632
fixes #633

+++ sdl/v1.go
@@ -35,8 +35,8 @@ type v1Service struct {
 }
 
 type v1Expose struct {
-	Port   uint32
-	As     uint32
+	Port   uint16
+	As     uint16
 	Proto  string       `yaml:",omitempty"`
 	To     []v1ExposeTo `yaml:",omitempty"`
 	Accept v1Accept
